Python/requirements.txt
# 基本ライブラリ
requests
beautifulsoup4
python-dotenv

# Google API関連
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread

# 日付・時間操作
pytz

# ファイル操作
Pillow

# スクレイピング
playwright

# システム監視
psutil
GPUtil

# OCR
pytesseract

# 機械学習・自動化
pyautogui
opencv-python

# データ処理
pandas
numpy

# ウェブソケット
websocket-client

# その他
PyDrive
moviepy
yt-dlp